Renamed packages: com.organization.mvcproject.control and com.organization.mvcproject.model to follow naming conventions
