+++ mgl-webapp/src/main/java/com/organization/mvcproject/controller/MGLTask1Controller.java
@@ -1,4 +1,3 @@
-//TODO 1.0   package naming convention, improve package declaration
 package com.organization.mvcproject.controller;
 
 import java.util.List;
@@ -14,9 +13,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.organization.mvcproject.MGL_Task1.model.Game;
-import com.organization.mvcproject.MGL_Task1.model.Review;
-import com.organization.mvcproject.MGL_Task1.service.GameService;
+import com.organization.mvcproject.model.Game;
+import com.organization.mvcproject.model.Review;
+import com.organization.mvcproject.service.GameService;
 
 @Controller
 public class MGLTask1Controller {
