+++ mgl-webapp/src/main/java/com/organization/mvcproject/model/Company.java
@@ -1,5 +1,4 @@
-
-package com.organization.mvcproject.MGL_Task1.model;
+package com.organization.mvcproject.model;
 
 import java.util.ArrayList;
 import java.util.List;
