+++ mgl-webapp/src/main/java/com/organization/mvcproject/model/Game.java
@@ -1,4 +1,4 @@
-package com.organization.mvcproject.MGL_Task1.model;
+package com.organization.mvcproject.model;
 
 import org.springframework.stereotype.Component;
